+++ plugins/benq_dvdrom/qscan_plugin.cpp
@@ -98,7 +98,7 @@ int  scan_benqrom::start_test(unsigned int itest, long ilba, int &speed)
 	return r;
 }
 
-int  scan_benqrom::scan_block(void *data, long *ilba)
+int  scan_benqrom::scan_block(void *data, uint32_t *ilba)
 {
 	int r=-1;
 	switch (test) {
